增加子设备上下线 调整文件结构

+++ model/interface.go
@@ -3,7 +3,7 @@ package model
 // Conn conn接口
 type Conn interface {
 	// Publish will publish a message with the specified QoS and content
-	Publish(topic string, qos int, payload interface{}) error
+	Publish(topic string, qos byte, payload interface{}) error
 	UnderlyingClient() interface{}
 	Subscribe(topic string, streamFunc ProcDownStreamFunc) error
 	ContainerOf() *Manager
